Solution to Check Profit or Loss

+++ Check Profit or Loss@@6788dd63490637cfc44e1b98/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    float cost price,selling price;
-    scanf("%f %f",&cost price,&selling price)
+    int cost price,selling price;
+    scanf("%d %d",&cost price,&selling price)
     if(selling price>cost price){
         printf("Profit");
     }
